pyproject.toml
[project]
name = "nonebot-plugin-elf-rss-next"
version = "0.1.0"
description = "RSS订阅机器人“ELF_RSS”的独立插件版本"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "arrow>=1.3.0",
    "async-timeout>=5.0.1",
    "asyncache>=0.3.1",
    "deep-translator>=1.11.4",
    "emoji>=2.14.1",
    "feedparser>=6.0.11",
    "imagehash>=4.3.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.59.4",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot2>=2.4.3",
    "pillow>=11.3.0",
    "pyquery>=2.0.1",
    "tenacity>=9.1.2",
    "tinydb>=4.8.2",
    "yarl>=1.20.1",
]
